refactor(tui): split monolithic src/tui.rs into specialized modules
